addressing information fix
Change dictionaries used by addressing information classes to always contain every attribute - to make it compatible and easily replaced by other addressing information types.
