+++ tests/software_tests/transport_interface/test_can_transport_interface.py
@@ -1,33 +1,31 @@
 import pytest
 from mock import MagicMock, Mock, patch
 
-from uds.transport_interface.abstract_can_transport_interface import AbstractCanTransportInterface, \
-    AbstractCanAddressingInformation
+from uds.transport_interface.can_transport_interface import AbstractCanTransportInterface, PyCanTransportInterface, \
+    AbstractCanAddressingInformation, BusABC
+from uds.can import CanAddressingInformation, CanAddressingFormat
 
 
-@pytest.mark.skip
+SCRIPT_LOCATION = "uds.transport_interface.can_transport_interface"
+
+
 class TestAbstractCanTransportInterface:
     """Unit tests for `AbstractCanTransportInterface` class."""
-
-    SCRIPT_LOCATION = "uds.transport_interface.abstract_can_transport_interface"
 
     def setup(self):
         self.mock_can_transport_interface = Mock(spec=AbstractCanTransportInterface)
         # patching
         self._patcher_abstract_transport_interface_init \
-            = patch(f"{self.SCRIPT_LOCATION}.AbstractTransportInterface.__init__")
+            = patch(f"{SCRIPT_LOCATION}.AbstractTransportInterface.__init__")
         self.mock_abstract_transport_interface_init = self._patcher_abstract_transport_interface_init.start()
-        self._patcher_can_segmenter_class = patch(f"{self.SCRIPT_LOCATION}.CanSegmenter")
+        self._patcher_can_segmenter_class = patch(f"{SCRIPT_LOCATION}.CanSegmenter")
         self.mock_can_segmenter_class = self._patcher_can_segmenter_class.start()
-        self._patcher_can_packet_class = patch(f"{self.SCRIPT_LOCATION}.CanPacket")
-        self.mock_can_packet_class = self._patcher_can_packet_class.start()
-        self._patcher_warn = patch(f"{self.SCRIPT_LOCATION}.warn")
+        self._patcher_warn = patch(f"{SCRIPT_LOCATION}.warn")
         self.mock_warn = self._patcher_warn.start()
 
     def teardown(self):
         self._patcher_abstract_transport_interface_init.stop()
         self._patcher_can_segmenter_class.stop()
-        self._patcher_can_packet_class.stop()
         self._patcher_warn.stop()
 
     # __init__
@@ -47,9 +45,8 @@ class TestAbstractCanTransportInterface:
         mock_isinstance.assert_called_once_with(addressing_information, AbstractCanAddressingInformation)
 
     @pytest.mark.parametrize("can_bus_manager, addressing_information", [
-        ("can_bus_manager", Mock(tx_packets_physical_ai={"addressing_format": Mock(),
-                                                         "addressing_type": Mock()})),
-        (Mock(), Mock(tx_packets_physical_ai={"arg1": 1, "arg2": 2})),
+        ("can_bus_manager", "some addressing information"),
+        (Mock(), Mock()),
     ])
     @patch(f"{SCRIPT_LOCATION}.isinstance")
     def test_init__valid_mandatory_args(self, mock_isinstance,
@@ -60,10 +57,7 @@ class TestAbstractCanTransportInterface:
                                                addressing_information=addressing_information)
         mock_isinstance.assert_called_once_with(addressing_information, AbstractCanAddressingInformation)
         self.mock_abstract_transport_interface_init.assert_called_once_with(bus_manager=can_bus_manager)
-        self.mock_can_segmenter_class.assert_called_once_with(
-            addressing_format=addressing_information.addressing_format,
-            physical_ai=addressing_information.tx_packets_physical_ai,
-            functional_ai=addressing_information.tx_packets_functional_ai)
+        self.mock_can_segmenter_class.assert_called_once_with(addressing_information=addressing_information)
         assert self.mock_can_transport_interface._AbstractCanTransportInterface__addressing_information \
                == addressing_information
         assert self.mock_can_transport_interface._AbstractCanTransportInterface__segmenter \
@@ -76,9 +70,8 @@ class TestAbstractCanTransportInterface:
         assert self.mock_can_transport_interface.n_cr_timeout == self.mock_can_transport_interface.N_CR_TIMEOUT
 
     @pytest.mark.parametrize("can_bus_manager, addressing_information", [
-        ("can_bus_manager", Mock(tx_packets_physical_ai={"addressing_format": Mock(),
-                                                         "addressing_type": Mock()})),
-        (Mock(), Mock(tx_packets_physical_ai={"arg1": 1, "arg2": 2})),
+        ("can_bus_manager", "addressing_information"),
+        (Mock(), Mock()),
     ])
     @pytest.mark.parametrize("n_as_timeout, n_ar_timeout, n_bs_timeout, n_br, n_cs, n_cr_timeout, "
                              "dlc, use_data_optimization, filler_byte", [
@@ -107,9 +100,7 @@ class TestAbstractCanTransportInterface:
         mock_isinstance.assert_called_once_with(addressing_information, AbstractCanAddressingInformation)
         self.mock_abstract_transport_interface_init.assert_called_once_with(bus_manager=can_bus_manager)
         self.mock_can_segmenter_class.assert_called_once_with(
-            addressing_format=addressing_information.addressing_format,
-            physical_ai=addressing_information.tx_packets_physical_ai,
-            functional_ai=addressing_information.tx_packets_functional_ai,
+            addressing_information=addressing_information,
             dlc=dlc,
             use_data_optimization=use_data_optimization,
             filler_byte=filler_byte)
@@ -500,3 +491,114 @@ class TestAbstractCanTransportInterface:
     def test_filler_byte__set(self, value):
         AbstractCanTransportInterface.filler_byte.fset(self.mock_can_transport_interface, value)
         assert self.mock_can_transport_interface.segmenter.filler_byte == value
+
+
+class TestPyCanTransportInterface:
+    """Unit tests for `PyCanTransportInterface` class."""
+
+    def setup(self):
+        self.mock_can_transport_interface = Mock(spec=PyCanTransportInterface)
+        # patching
+        self._patcher_abstract_can_ti_init = patch(f"{SCRIPT_LOCATION}.AbstractCanTransportInterface.__init__")
+        self.mock_abstract_can_ti_init = self._patcher_abstract_can_ti_init.start()
+
+    def teardown(self):
+        self._patcher_abstract_can_ti_init.stop()
+
+    # __init__
+
+    @pytest.mark.parametrize("can_bus_manager, addressing_information", [
+        ("can_bus_manager", "addressing_information"),
+        (Mock(), Mock())
+    ])
+    def test_init__default_args(self, can_bus_manager, addressing_information):
+        PyCanTransportInterface.__init__(self=self.mock_can_transport_interface,
+                                         can_bus_manager=can_bus_manager,
+                                         addressing_information=addressing_information)
+        self.mock_abstract_can_ti_init.assert_called_once_with(
+            can_bus_manager=can_bus_manager,
+            addressing_information=addressing_information)
+        assert self.mock_can_transport_interface._PyCanTransportInterface__n_as_measured is None
+        assert self.mock_can_transport_interface._PyCanTransportInterface__n_ar_measured is None
+        assert self.mock_can_transport_interface._PyCanTransportInterface__n_bs_measured is None
+        assert self.mock_can_transport_interface._PyCanTransportInterface__n_cr_measured is None
+
+    @pytest.mark.parametrize("can_bus_manager, addressing_information", [
+        ("can_bus_manager", "addressing_information"),
+        (Mock(), Mock())
+    ])
+    @pytest.mark.parametrize("kwargs", [
+        {"a": Mock(), "b": Mock()},
+        {"param1": Mock(), "param2": Mock(), "something_else": Mock()},
+    ])
+    def test_init__all_args(self, can_bus_manager, addressing_information, kwargs):
+        PyCanTransportInterface.__init__(self=self.mock_can_transport_interface,
+                                         can_bus_manager=can_bus_manager,
+                                         addressing_information=addressing_information,
+                                         **kwargs)
+        self.mock_abstract_can_ti_init.assert_called_once_with(can_bus_manager=can_bus_manager,
+                                                               addressing_information=addressing_information,
+                                                               **kwargs)
+        assert self.mock_can_transport_interface._PyCanTransportInterface__n_as_measured is None
+        assert self.mock_can_transport_interface._PyCanTransportInterface__n_ar_measured is None
+        assert self.mock_can_transport_interface._PyCanTransportInterface__n_bs_measured is None
+        assert self.mock_can_transport_interface._PyCanTransportInterface__n_cr_measured is None
+
+    # n_as_measured
+
+    @pytest.mark.parametrize("value", ["something", Mock()])
+    def test_n_as_measured(self, value):
+        self.mock_can_transport_interface._PyCanTransportInterface__n_as_measured = value
+        assert PyCanTransportInterface.n_as_measured.fget(self.mock_can_transport_interface) == value
+
+    # n_as_measured
+
+    @pytest.mark.parametrize("value", ["something", Mock()])
+    def test_n_ar_measured(self, value):
+        self.mock_can_transport_interface._PyCanTransportInterface__n_ar_measured = value
+        assert PyCanTransportInterface.n_ar_measured.fget(self.mock_can_transport_interface) == value
+
+    # n_bs_measured
+
+    @pytest.mark.parametrize("value", ["something", Mock()])
+    def test_n_bs_measured(self, value):
+        self.mock_can_transport_interface._PyCanTransportInterface__n_bs_measured = value
+        assert PyCanTransportInterface.n_bs_measured.fget(self.mock_can_transport_interface) == value
+
+    # n_cr_measured
+
+    @pytest.mark.parametrize("value", ["something", Mock()])
+    def test_n_cr_measured(self, value):
+        self.mock_can_transport_interface._PyCanTransportInterface__n_cr_measured = value
+        assert PyCanTransportInterface.n_cr_measured.fget(self.mock_can_transport_interface) == value
+
+    # is_supported_bus_manager
+
+    @pytest.mark.parametrize("value", ["something", Mock()])
+    @patch(f"{SCRIPT_LOCATION}.isinstance")
+    def test_is_supported_bus_manager(self, mock_isinstance, value):
+        assert PyCanTransportInterface.is_supported_bus_manager(value) == mock_isinstance.return_value
+        mock_isinstance.assert_called_once_with(value, BusABC)
+
+
+class TestPyCanTransportInterfaceIntegration:
+    """Integration tests for `PyCanTransportInterface` class."""
+
+    @pytest.mark.parametrize("init_kwargs", [
+        {
+            "can_bus_manager": Mock(spec=BusABC),
+            "addressing_information": CanAddressingInformation(
+                addressing_format=CanAddressingFormat.NORMAL_11BIT_ADDRESSING,
+                rx_physical={"can_id": 0x641},
+                tx_physical={"can_id": 0x642},
+                rx_functional={"can_id": 0x6FE},
+                tx_functional={"can_id": 0x6FF},
+            ),
+        }
+    ])
+    def test_init(self, init_kwargs):
+        py_can_ti = PyCanTransportInterface(**init_kwargs)
+        assert py_can_ti.n_as_measured is None
+        assert py_can_ti.n_ar_measured is None
+        assert py_can_ti.n_bs_measured is None
+        assert py_can_ti.n_cr_measured is None
