+++ uds/can/normal_addressing_information.py
@@ -8,7 +8,7 @@ from uds.utilities import InconsistentArgumentsError, UnusedArgumentError, valid
 from uds.transmission_attributes import AddressingType, AddressingTypeAlias
 from .addressing_format import CanAddressingFormat, CanAddressingFormatAlias
 from .frame_fields import CanIdHandler
-from .abstract_addressing_information import AbstractCanAddressingInformation
+from .abstract_addressing_information import AbstractCanAddressingInformation, PacketAIParamsAlias
 
 
 class Normal11BitCanAddressingInformation(AbstractCanAddressingInformation):
@@ -29,7 +29,7 @@ class Normal11BitCanAddressingInformation(AbstractCanAddressingInformation):
                            target_address: Optional[int] = None,
                            source_address: Optional[int] = None,
                            address_extension: Optional[int] = None
-                           ) -> AbstractCanAddressingInformation.PacketAIParamsAlias:
+                           ) -> PacketAIParamsAlias:
         """
         Validate Addressing Information parameters of a CAN packet that uses Normal 11-bit Addressing format.
 
@@ -52,11 +52,14 @@ class Normal11BitCanAddressingInformation(AbstractCanAddressingInformation):
         if not CanIdHandler.is_normal_11bit_addressed_can_id(can_id):  # type: ignore
             raise InconsistentArgumentsError(f"Provided value of CAN ID is not compatible with "
                                              f"Normal 11-bit Addressing Format. Actual value: {can_id}")
-        return {
-            cls.ADDRESSING_FORMAT_NAME: CanAddressingFormat.NORMAL_11BIT_ADDRESSING,  # type: ignore
-            cls.ADDRESSING_TYPE_NAME: addressing_type,
-            cls.CAN_ID_NAME: can_id
-        }
+        return PacketAIParamsAlias(
+            addressing_format=CanAddressingFormat.NORMAL_11BIT_ADDRESSING,
+            addressing_type=addressing_type,
+            can_id=can_id,  # type: ignore
+            target_address=target_address,
+            source_address=source_address,
+            address_extension=address_extension
+        )
 
 
 class NormalFixedCanAddressingInformation(AbstractCanAddressingInformation):
@@ -77,7 +80,7 @@ class NormalFixedCanAddressingInformation(AbstractCanAddressingInformation):
                            target_address: Optional[int] = None,
                            source_address: Optional[int] = None,
                            address_extension: Optional[int] = None
-                           ) -> AbstractCanAddressingInformation.PacketAIParamsAlias:
+                           ) -> PacketAIParamsAlias:
         """
         Validate Addressing Information parameters of a CAN packet that uses Normal Fixed Addressing format.
 
@@ -109,13 +112,14 @@ class NormalFixedCanAddressingInformation(AbstractCanAddressingInformation):
                 addressing_type=addressing_type,
                 target_address=target_address,  # type: ignore
                 source_address=source_address)  # type: ignore
-            return {
-                cls.ADDRESSING_FORMAT_NAME: CanAddressingFormat.NORMAL_FIXED_ADDRESSING,  # type: ignore
-                cls.ADDRESSING_TYPE_NAME: addressing_type,
-                cls.CAN_ID_NAME: encoded_can_id,
-                cls.TARGET_ADDRESS_NAME: target_address,
-                cls.SOURCE_ADDRESS_NAME: source_address,
-            }
+            return PacketAIParamsAlias(
+                addressing_format=CanAddressingFormat.NORMAL_FIXED_ADDRESSING,
+                addressing_type=addressing_type,
+                can_id=encoded_can_id,
+                target_address=target_address,
+                source_address=source_address,
+                address_extension=address_extension
+            )
         decoded_info = CanIdHandler.decode_normal_fixed_addressed_can_id(can_id)
         if addressing_type != decoded_info[CanIdHandler.ADDRESSING_TYPE_NAME]:  # type: ignore
             raise InconsistentArgumentsError(f"Provided value of CAN ID is not compatible with Addressing Type."
@@ -126,10 +130,11 @@ class NormalFixedCanAddressingInformation(AbstractCanAddressingInformation):
         if source_address not in (decoded_info[CanIdHandler.SOURCE_ADDRESS_NAME], None):  # type: ignore
             raise InconsistentArgumentsError(f"Provided value of CAN ID is not compatible with Source Address."
                                              f"Actual values: can_id={can_id}, source_address={source_address}")
-        return {
-            cls.ADDRESSING_FORMAT_NAME: CanAddressingFormat.NORMAL_FIXED_ADDRESSING,  # type: ignore
-            cls.ADDRESSING_TYPE_NAME: addressing_type,
-            cls.CAN_ID_NAME: can_id,
-            cls.TARGET_ADDRESS_NAME: decoded_info[CanIdHandler.TARGET_ADDRESS_NAME],
-            cls.SOURCE_ADDRESS_NAME: decoded_info[CanIdHandler.SOURCE_ADDRESS_NAME],
-        }
+        return PacketAIParamsAlias(
+            addressing_format=CanAddressingFormat.NORMAL_FIXED_ADDRESSING,
+            addressing_type=addressing_type,
+            can_id=can_id,
+            target_address=decoded_info[CanIdHandler.TARGET_ADDRESS_NAME],  # type: ignore
+            source_address=decoded_info[CanIdHandler.SOURCE_ADDRESS_NAME],  # type: ignore
+            address_extension=address_extension
+        )
